deps: depend on node-sqlite3 from git
The commit is marked as 3.0.11 but has not yet been tagged and released.

The only change in it is the upgrade to nan@2, which allows building against iojs-v3 and node-v4.

See https://github.com/mapbox/node-sqlite3/commit/5a2939d0f1a0929c0541a2a8dce327dc38d26421

Note that this does not fix the failing tests, only the failing install.